Backend/ai_service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
sentence-transformers==2.2.2
chromadb==0.4.15
PyPDF2==3.0.1
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0